Update 5
Due to an engine error I had to make a new game project in Unity.
However I fixed player rotating, added enemy and AI movement, fixed an
error the would cause the player to not be able to move and change the
orientation of the player as well.  Over this was a big bug fix and re
write of lost data.